requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.0
defusedxml==0.7.1
Django==4.0.4
django-allauth==0.50.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
drf-spectacular==0.22.1
idna==3.3
importlib-resources==5.7.1
inflection==0.5.1
jsonschema==4.4.0
oauthlib==3.2.0
paho-mqtt==1.6.1
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.1
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.8.0
